feature : 프리셋 도시에 대해서 날씨 리스트 불러오기

+++ WantedPreOnboarding/Models/City.swift
@@ -28,6 +28,6 @@ struct City: Decodable {
 }
 
 struct Coordinate: Decodable {
-    let lon: Float
-    let lat: Float
+    let lon: Double
+    let lat: Double
 }
